+++ test/phoenix/integration/long_poll_test.exs
@@ -213,7 +213,7 @@ defmodule Phoenix.Integration.LongPollTest do
     end
 
     test "#{@mode}: publishing events" do
-      Phoenix.PubSub.subscribe(__MODULE__, self, "rooms:lobby")
+      Phoenix.PubSub.subscribe(__MODULE__, "rooms:lobby")
       session = join("/ws", "rooms:lobby", @mode)
 
       # Publish successfuly
@@ -237,7 +237,7 @@ defmodule Phoenix.Integration.LongPollTest do
 
       # Publish unauthorized event
       capture_log fn ->
-        Phoenix.PubSub.subscribe(__MODULE__, self, "rooms:private-room")
+        Phoenix.PubSub.subscribe(__MODULE__, "rooms:private-room")
         resp = poll :post, "/ws", session, %{
           "topic" => "rooms:private-room",
           "event" => "new_msg",