+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.2.0-dev
 
+* Deprecations
+  * [Endpoint] Generated `subscribe/3` and `unsubscibe/2` clauses have been deprecated in favor of `subscribe/2` and `unsubscribe/1` which uses the caller's pid.
+  * [PubSub] `Phoenix.PubSub.subscribe/3` and `Phoenix.PubSub.unsubscibe/2` have been deprecated in favor of `subscribe/2` and `unsubscribe/1` which uses the caller's pid.
+
 ## 1.1.4 (2016-1-25)
 
 * Enhancements